refactor(bzm2): extract hash thread assignment tracker
